requirements.txt
flask>=3.0
pdfplumber>=0.10
pytesseract>=0.3
Pillow>=10.0
spacy>=3.7
marshmallow>=3.20
python-dotenv>=1.0
